Joined and flat version of EDI data package knb-lter-hfr.118.33

A fully joined and flat version of EDI data package knb-lter-hfr.118.33 (Ant Assemblages in Hemlock Removal Experiment at Harvard Forest since 2003) with all relevant ecocomDP L1 identifiers and content added. Use this dataset as an input to the L0_flat argument of the "create" functions.

ants_L0_flat

Format

A data frame with 2931 rows and 45 variables:

datetime

dates

block

block

plot

plot number

treatment

treatment type

moose.cage

location of grid with respect to moose exclosure

trap.type

trap type

trap.num

applies only to pitfall cups

subfamily

ant subfamily

hl

head length. We used trait definitions from Del Toro et al. (2015) and filled in missing species' data with information from Ellison et al.

rel

eye length relative to body size

rll

femur length relative to body size

colony.size

size of colony for each species

feeding.preference

feeding preference for each species

nest.substrate

nest substrate

primary.habitat

primary habitat

secondary.habitat

secondary habitat associations

seed.disperser

whether or not a seed dispersing species

slavemaker.sp

whether or not a slavemaking species

behavior

classifications based on behavioral interactions with other ants

biogeographic.affinity

biogeographic affinity based on available occurrence records

source

where trait information was found. Full citations for literature are as follows: Del Toro, I., R.R. Silva, and A.M. Ellison. 2015. Predicated impacts of climatic change on ant functional diversity and distributions in eastern North American forests. Diversity and Distributions 21:781-791; Ellison, A.M., N.J. Gotelli, G. Alpert, and E.J. Farnsworth. 2012. A field guide to the ants of New England. Yale University Press, New Haven, Connecticut, USA.

unit_hl

units for "hl" variable

unit_rel

units for "rel" variable

unit_rll

units for "rll" variable

variable_name

variables of the primary observation table

value

values of variable_name

unit

units of variable_name

observation_id

the observation id

location_id

the location id

event_id

the event id

latitude

approximate latitude of study area

longitude

approximate longitude of study area

elevation

approximate elevation of study area

taxon_name

name of organism

taxon_id

the taxon id

taxon_rank

the taxon rank

authority_system

the authority system taxon_name was resolved to

authority_taxon_id

the id of taxon_name in authority_system

package_id

the identifier of this ecocomDP dataset

original_package_id

the identifier of the source dataset

length_of_survey_years

number of years the survey has been ongoing

number_of_years_sampled

number of years during the survey that samples were taken

std_dev_interval_betw_years

the standard deviation between surveys in years

max_num_taxa

number of unique taxa in this dataset

geo_extent_bounding_box_m2

the study area in meters squared

Source

https://portal.edirepository.org/nis/mapbrowse?scope=knb-lter-hfr&identifier=118&revision=33