Make EML metadata for a DWcA occurrence from an ecocomDP data package

Make EML metadata for a DWcA occurrence from an ecocomDP data package

make_eml_dwca(
  path,
  core_name,
  source_id,
  derived_id,
  user.id,
  user.domain,
  url = NULL
)

Arguments

path
(character) Path to the directory containing ecocomDP data tables, conversion scripts, and where EML metadata will be written. This path, when defined on a web server, also serves as the publicly accessible URL from which the data objects can be downloaded.

core_name
(character) The Darwin Core central table of the package. Can be: "occurence" (occurrence core) or "event" (event core).

source_id
(character) ID of an ecocomDP data package. Only EDI Data Repository package IDs are currently supported.

derived_id
(character) Identifier of the DwC-A dataset being created.

user.id
(character; optional) Repository user identifier. If more than one, then enter as a vector of character strings (e.g. c("user_id_1", "user_id_2")). user.id sets the /eml/access/principal element for all user.domain except "KNB", "ADC", and if user.domain = NULL.

user.domain
(character; optional) Repository domain associated with user.id. Currently supported values are "EDI" (Environmental Data Initiative), "LTER" (Long-Term Ecological Research Network), "KNB" (The Knowledge Network for Biocomplexity), "ADC" (The Arctic Data Center). If you'd like your system supported please contact maintainers of the ecocomDP R package. If using more than one user.domain, then enter as a vector of character strings (e.g. c("user_domain_1", "user_domain_2")) in the same order as corresponding user.id. If user.domain is missing then a default value "unknown" is assigned. user.domain sets the EML header "system" attribute and for all user.domain, except "KNB" and "ADC", sets the /eml/access/principal element attributes and values.

url
(character) URL to the publicly accessible directory containing DwC-A tables and meta.xml. This argument supports direct download of the data entities by a data repository and is used within the scope of the ecocomDP project for automated revision and upload of ecocomDP data packages and derived products.

repository
(character) Data repository in which package.id resides and associated with environment. Currently supported repositories are: "EDI" (Environmental Data Initiative). Requests for support of other repositories can be made via ecocomDP GitHub issues. Default is "EDI".

Value

An EML metadata record for the DWcA table defined by data.table.

Details

This function creates an EML record for an Darwin Core Archive record (DwC-A) combining metadata from the parent data package and boiler-plate metadata describing the DwC-A tables. Changes to the parent EML include:

<access> Adds the user.id to the list of principals granted read and write access to the DwC-A data package this EML describes.

<title> Appends "Darwin Core Archive: " to the title.

<pubDate> Adds the date when this EML record is created.

<abstract> Adds a note that this is a derived data package in a DwC-A format.

<keywordSet> Essential Biodiversity Variables: "Population Abundance" and Darwin Core Terms: "BasisofRecord: HumanObservation", "Occurrence: OrganismQuantity", "Taxon: ScientificName".

<intellectualRights> Keeps intact the intellectual rights license of the parent data package, or replaces it with "CCO" (https://creativecommons.org/publicdomain/zero/1.0/legalcode).

<methodStep> Adds a note that this data package was created by methods within the ecocomDP R package and adds provenance metadata noting that this is a derived data and describing where the parent data package can be accessed.

<dataTables> Replaces the parent data package data tables metadata with boiler-plate metadata for the DwC-A tables.

<otherEntity> Describes the meta.xml accompanying each DwC-A. Any other entities listed in the parent EML are removed.

Examples

if (FALSE) {

}