Map NEON data products to ecocomDP

Pull files from the NEON API by data product, merge data for each table, and convert to ecocomDP format. Please see neonUtilities::loadByProduct for more information.

map_neon.ecocomdp.10003.001.001(
  neon.data.list,
  neon.data.product.id = "DP1.10003.001",
  ...
)

map_neon.ecocomdp.10022.001.001(
  neon.data.list,
  neon.data.product.id = "DP1.10022.001",
  ...
)

map_neon.ecocomdp.10022.001.002(
  neon.data.list,
  neon.data.product.id = "DP1.10022.001",
  ...
)

map_neon.ecocomdp.10043.001.001(
  neon.data.list,
  neon.data.product.id = "DP1.10043.001",
  ...
)

map_neon.ecocomdp.10058.001.001(
  neon.data.list,
  neon.data.product.id = "DP1.10058.001",
  ...
)

map_neon.ecocomdp.10072.001.001(
  neon.data.list,
  neon.data.product.id = "DP1.10072.001",
  ...
)

map_neon.ecocomdp.10092.001.001(
  neon.data.list,
  neon.data.product.id = "DP1.10092.001",
  ...
)

map_neon.ecocomdp.10093.001.001(
  neon.data.list,
  neon.data.product.id = "DP1.10093.001",
  ...
)

map_neon.ecocomdp.20107.001.001(
  neon.data.list,
  neon.data.product.id = "DP1.20107.001",
  ...
)

map_neon.ecocomdp.20120.001.001(
  neon.data.list,
  neon.data.product.id = "DP1.20120.001",
  ...
)

map_neon.ecocomdp.20166.001.001(
  neon.data.list,
  neon.data.product.id = "DP1.20166.001",
  ...
)

map_neon.ecocomdp.20219.001.001(
  neon.data.list,
  neon.data.product.id = "DP1.20219.001",
  ...
)

map_neon_data_to_ecocomDP(
  id,
  neon.data.save.dir = NULL,
  neon.data.read.path = NULL,
  neon.data.list = NULL,
  ...
)

Arguments

neon.data.list
(list) A list of stacked NEON data tables to be mapped to ecocomDP. Must match the chosen mapping method in 'id'

...
Additional arguments passed to neonUtilities::loadByProduct().

id
(character) Identifier of dataset to read. Identifiers are listed in the "id" column of the search_data() output. e.g., neon.ecocomdp.20166.001.001 is the id for "ALGAE Periphyton, seston, and phytoplankton collection" converted to the ecocomDP format from the NEON source data product DP1.20166.001.

neon.data.save.dir
(character) Directory to save NEON source data that has been downloaded via neonUtilities::loadByProduct()

neon.data.read.path
(character) Path to read in an RDS file of 'stacked NEON data' from neonUtilities::loadByProduct()

site
Either the string 'all', meaning all available sites, or a character vector of 4-letter NEON site codes, e.g. c('ONAQ','RMNP'). Defaults to all.

startdate
Either NA, meaning all available dates, or a character vector in the form YYYY-MM, e.g. 2017-01. Defaults to NA.

enddate
Either NA, meaning all available dates, or a character vector in the form YYYY-MM, e.g. 2017-01. Defaults to NA.

check.size
T or F, should the user approve the total file size before downloading? Defaults to T. When working in batch mode, or other non-interactive workflow, use check.size=F.

nCores
The number of cores to parallelize the stacking procedure. By default it is set to a single core.

forceParallel
If the data volume to be processed does not meet minimum requirements to run in parallel, this overrides. Set to FALSE as default.

token
User specific API token (generated within neon.datascience user accounts)

Value

Returns a named list, including:

metadata

A list of information about the NEON data product returned by neonUtilities::getProductInfo.

tables

A list of data.frames following the ecocomDP format.

tables$location

A table, which has the lat long for each NEON site (location_id) included in the data set

tables$taxon

A table, which describes each taxon_id that appears in the data set

tables$observation

A table, with a record for each observation_id that appears in the data set

Details

All available data meeting the query criteria will be downloaded. Most data products are collected at only a subset of sites, and dates of collection vary. Consult the NEON data portal for sampling details. Dates are specified only to the month because NEON data are provided in monthly packages. Any month included in the search criteria will be included in the download. Start and end date are inclusive.

Functions

map_neon.ecocomdp.10003.001.001: This method will retrieve count data for BIRD taxa from neon.data.product.id DP1.10003.001 from the NEON data portal and map to the ecocomDP format

map_neon.ecocomdp.10022.001.001: This method will retrieve density data for BEETLE from neon.data.product.id DP1.10022.001 from the NEON data portal and map to the ecocomDP

map_neon.ecocomdp.10022.001.002: This method will retrieve count data for HERPTILE taxa from neon.data.product.id DP1.10022.001 from the NEON data portal and map to the ecocomDP format

map_neon.ecocomdp.10043.001.001: This method will retrieve density data for MOSQUITO from neon.data.product.id DP1.10043.001 from the NEON data portal and map to the ecocomDP

map_neon.ecocomdp.10058.001.001: This method will retrieve percent cover data for PLANT taxa from neon.data.product.id DP1.10058.001 from the NEON data portal and map to the ecocomDP format

map_neon.ecocomdp.10072.001.001: This method will retrieve count data for SMALL_MAMMAL taxa from neon.data.product.id DP1.10072.001 from the NEON data portal and map to the ecocomDP format

map_neon.ecocomdp.10092.001.001: This method will retrieve positivity data for TICK_PATHOGEN taxa from neon.data.product.id DP1.10092.001 from the NEON data portal and map to the ecocomDP format

map_neon.ecocomdp.10093.001.001: This method will retrieve count data for TICK taxa from neon.data.product.id DP1.10093.001 from the NEON data portal and map to the ecocomDP format

map_neon.ecocomdp.20107.001.001: This method will retrieve count data for FISH taxa from neon.data.product.id DP1.20107.001 from the NEON data portal and map to the ecocomDP format

map_neon.ecocomdp.20120.001.001: This method will retrieve density data for MACROINVERTEBRATE from neon.data.product.id DP1.20120.001 from the NEON data portal and map to the ecocomDP format

map_neon.ecocomdp.20166.001.001: This method will retrieve density data for ALGAE from neon.data.product.id DP1.20166.001 (Periphyton, seston, and phytoplankton collection) from the NEON data portal and map to the ecocomDP format

map_neon.ecocomdp.20219.001.001: This method will retrieve count data for ZOOPLANKTON taxa from neon.data.product.id DP1.20219.001 from the NEON data portal and map to the ecocomDP format

References

License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007

Author

Daijiang Li

Kari Norman kari.norman@berkeley.edu

Matt Helmus mrhelmus@temple.edu

Natalie Robinson nrobinson@battelleecology.org

Michael Just

Marta Jarzyna

Melissa Chen

Matt Bitters

Wynne Moss

Brendan Hobart

Thilina Surasinghe tsurasinghe@bridgew.edu

Stephanie Parker sparker@battelleecology.org

Lara Janson

Eric R Sokol esokol@battelleecology.org

Ruvi Jaimes

Examples

if (FALSE) {
# Retrieve and map NEON ALGAE data to ecocomDP format
my_result <- map_neon_data_to_ecocomDP(
  id = "neon.ecocomdp.20166.001.001",
  site = c("MAYF", "PRIN"),
  startdate = "2016-1",
  enddate = "2018-11",
  check.size = FALSE)
}

if (FALSE) {
}