Check for file name errors

This function ensures that ecocomDP file names follow the naming convention (i.e. studyName_ecocomDPTableName.ext, e.g. gleon_chloride_observation.csv).

validate_table_names(data.path = NULL)

Arguments

data.path
(character) Path to the directory containing ecocomDP tables.

Value

(character) If table names are valid, then the corresponding table names are returned. If table names are invalid, then an error message is returned.