Read an ecocomDP dataset

Read an ecocomDP dataset

read_data(
  id = NULL,
  parse_datetime = TRUE,
  unique_keys = FALSE,
  site = "all",
  startdate = NA,
  enddate = NA,
  package = "basic",
  check.size = FALSE,
  nCores = 1,
  forceParallel = FALSE,
  token = NA,
  neon.data.save.dir = NULL,
  neon.data.read.path = NULL,
  ...,
  from = NULL
)

Arguments

id
(character) Identifier of dataset to read. Identifiers are listed in the "id" column of the search_data() output. Older versions of datasets can be read, but a warning is issued.

parse_datetime
(logical) Parse datetime values if TRUE, otherwise return as character strings.

unique_keys
(logical) Whether to create globally unique primary keys (and associated foreign keys). Useful in maintaining referential integrity when working with multiple datasets. If TRUE, id is appended to each table's primary key and associated foreign key. Default is FALSE.

site
(character) For NEON data, a character vector of site codes to filter data on. Sites are listed in the "sites" column of the search_data() output. Defaults to "all", meaning all sites.

startdate
(character) For NEON data, the start date to filter on in the form YYYY-MM. Defaults to NA, meaning all available dates.

enddate
(character) For NEON data, the end date to filter on in the form YYYY-MM. Defaults to NA, meaning all available dates.

package
(character) For NEON data, either 'basic' or 'expanded', indicating which data package to download. Defaults to basic.

check.size
(logical) For NEON data, should the user approve the total file size before downloading? Defaults to FALSE.

nCores
(integer) For NEON data, the number of cores to parallelize the stacking procedure. Defaults to 1.

forceParallel
(logical) For NEON data, if the data volume to be processed does not meet minimum requirements to run in parallel, this overrides. Defaults to FALSE.

token
(character) For NEON data, a user specific API token (generated within neon.datascience user accounts).

neon.data.save.dir
(character) For NEON data, an optional and experimental argument (i.e. may not be supported in future releases), indicating the directory where NEON source data should be saved upon download from the NEON API. Data are downloaded using neonUtilities::loadByProduct() and saved in this directory as an .rds file. The filename will follow the format <NEON data product ID>_<timestamp>.rds

neon.data.read.path
(character) For NEON data, an optional and experimental argument (i.e. may not be supported in future releases), defining a path to read in an .rds file of 'stacked NEON data' from neonUtilities::loadByProduct(). See details below for more information.

...
For NEON data, other arguments to neonUtilities::loadByProduct()

from
(character) Full path of file to be read (if .rds), or path to directory containing saved datasets (if .csv).

Value

(list) with the structure:

id - Dataset identifier

metadata - List of info about the dataset. NOTE: This object is underdevelopment and content may change in future releases.

tables - List of dataset tables as data.frames.

validation_issues - List of validation issues. If the dataset fails any validation checks, then descriptions of each issue are listed here.

Details

Validation checks are applied to each dataset ensuring it complies with the ecocomDP model. A warning is issued when any validation checks fail. All datasets are returned, even if they fail validation.

Column classes are coerced to those defined in the ecocomDP specification.

Validation happens each time files are read, from source APIs or local environments.

Details for read_data() function regarding NEON data: Using this function to read data with an id that begins with "neon.ecocomdp" will result in a query to download NEON data from the NEON Data Portal API using neonUtilities::loadByProduct(). If a query includes provisional data (or if you are not sure if the query includes provisional data), we recommend saving a copy of the data in the original format provided by NEON in addition to the derived ecocomDP data package. To do this, provide a directory path using the neon.data.read.path argument. For example, the query my_ecocomdp_data <- read_data(id = "neon.ecocomdp.10022.001.001", neon.data.save.dir = "my_neon_data") will download the data for NEON Data Product ID DP1.10022.001 (ground beetles in pitfall traps) and convert it to the ecocomDP data model. In doing so, a copy of the original NEON download will be saved in the directory "my_ neon_data with the filename "DP1.10022.001_<timestamp>.RDS" and the derived data package in the ecocomDP format will be stored in your R environment in an object named "my_ecocomdp_data". Further, if you wish to reload a previously downloaded NEON dataset into the ecocomDP format, you can do so using my_ecocomdp_data <- read_data(id = "neon.ecocomdp.10022.001.001", neon.data.read.path = "my_neon_data/DP1.10022.001_<timestamp>.RDS")

Provisional NEON data. Despite NEON's controlled data entry, at times, errors are found in published data; for example, an analytical lab may adjust its calibration curve and re-calculate past analyses, or field scientists may discover a past misidentification. In these cases, Level 0 data are edited and the data are re-processed to Level 1 and re-published. Published data files include a time stamp in the file name; a new time stamp indicates data have been re-published and may contain differences from previously published data. Data are subject to re-processing at any time during an initial provisional period; data releases are never re-processed. All records downloaded from the NEON API will have a "release" field. For any provisional record, the value of this field will be "PROVISIONAL", otherwise, this field will have a value indicating the version of the release to which the record belongs. More details can be found at https://www.neonscience.org/data-samples/data-management/data-revisions-releases.

Note

This function may not work between 01:00 - 03:00 UTC on Wednesdays due to regular maintenance of the EDI Data Repository.

Examples

if (FALSE) {
# Read from EDI
dataset <- read_data("edi.193.4")
# str(dataset)
}

if (FALSE) {
# Read from NEON (full dataset)
dataset <- read_data("neon.ecocomdp.20120.001.001")
}

if (FALSE) {
# Read from NEON with filters (partial dataset)
dataset <- read_data(
  id = "neon.ecocomdp.20120.001.001", 
  site = c("COMO", "LECO", "SUGG"),
  startdate = "2017-06", 
  enddate = "2019-09",
  check.size = FALSE)
}

if (FALSE) {
# Read with datetimes as character
dataset <- read_data("edi.193.4", parse_datetime = FALSE)
is.character(dataset$edi.193.4$tables$observation$datetime)
}

# Save a list of datasets for reading
datasets <- c(ants_L1, ants_L1, ants_L1)  # 3 of the same, w/different names
names(datasets) <- c("ds1", "ds2", "ds3")
mypath <- paste0(tempdir(), "/data")      # A place for saving
dir.create(mypath)
save_data(datasets, mypath)               # Save as .rds

#> Writing datasets.rds to /var/folders/_g/7fy1qf010hb6fwxhhj_kmvb40000gn/T//Rtmp3CgTHK/data

save_data(datasets, mypath, type = ".csv")# Save as .csv

#> Writing ds1, ds2, ds3 as .csv to /var/folders/_g/7fy1qf010hb6fwxhhj_kmvb40000gn/T//Rtmp3CgTHK/data

# Read from local .rds
datasets <- read_data(from = paste0(mypath, "/datasets.rds"))

# Read from local .csv
datasets <- read_data(from = mypath)

# A dataset is returned as a list of metadata, tables, and validation issues
# (if there are any). The dataset ID is assigned to the top level for 
# reference.
str(datasets[1])

#> List of 1
#>  $ ds1:List of 2
#>   ..$ metadata: NULL
#>   ..$ tables  :List of 8
#>   .. ..$ location             :'data.frame':	10 obs. of  6 variables:
#>   .. .. ..$ location_id       : chr [1:10] "a1" "a2" "1" "2" ...
#>   .. .. ..$ location_name     : chr [1:10] "block_Valley" "block_Ridge" "plot_1" "plot_2" ...
#>   .. .. ..$ latitude          : num [1:10] NA NA 42.5 42.5 42.5 ...
#>   .. .. ..$ longitude         : num [1:10] NA NA -72.2 -72.2 -72.2 ...
#>   .. .. ..$ elevation         : num [1:10] NA NA 220 220 220 220 220 220 220 220
#>   .. .. ..$ parent_location_id: chr [1:10] NA NA "a1" "a1" ...
#>   .. ..$ taxon                :'data.frame':	53 obs. of  5 variables:
#>   .. .. ..$ taxon_id          : chr [1:53] "1" "2" "3" "4" ...
#>   .. .. ..$ taxon_rank        : chr [1:53] "Species" "Species" "Species" "Species" ...
#>   .. .. ..$ taxon_name        : chr [1:53] "Aphaenogaster fulva" "Aphaenogaster picea" "Camponotus chromaiodes" "Camponotus herculeanus" ...
#>   .. .. ..$ authority_system  : chr [1:53] "ITIS" "ITIS" "ITIS" "ITIS" ...
#>   .. .. ..$ authority_taxon_id: chr [1:53] "578383" "578440" "575766" "575995" ...
#>   .. ..$ observation          :'data.frame':	2931 obs. of  9 variables:
#>   .. .. ..$ observation_id: chr [1:2931] "1" "2" "3" "4" ...
#>   .. .. ..$ event_id      : chr [1:2931] "1" "1" "1" "1" ...
#>   .. .. ..$ package_id    : chr [1:2931] "edi.193.4" "edi.193.4" "edi.193.4" "edi.193.4" ...
#>   .. .. ..$ location_id   : chr [1:2931] "1" "1" "1" "1" ...
#>   .. .. ..$ datetime      : Date[1:2931], format: "2003-06-01" ...
#>   .. .. ..$ taxon_id      : chr [1:2931] "2" "6" "2" "6" ...
#>   .. .. ..$ variable_name : chr [1:2931] "abundance" "abundance" "abundance" "abundance" ...
#>   .. .. ..$ value         : num [1:2931] 2 1 2 1 1 1 1 1 1 1 ...
#>   .. .. ..$ unit          : chr [1:2931] "number" "number" "number" "number" ...
#>   .. ..$ location_ancillary   :'data.frame':	8 obs. of  6 variables:
#>   .. .. ..$ location_ancillary_id: chr [1:8] "1" "2" "3" "4" ...
#>   .. .. ..$ location_id          : chr [1:8] "1" "2" "3" "4" ...
#>   .. .. ..$ datetime             : Date[1:8], format: NA ...
#>   .. .. ..$ variable_name        : chr [1:8] "treatment" "treatment" "treatment" "treatment" ...
#>   .. .. ..$ value                : chr [1:8] "Girdled" "Logged" "HemlockControl" "Logged" ...
#>   .. .. ..$ unit                 : chr [1:8] NA NA NA NA ...
#>   .. ..$ taxon_ancillary      :'data.frame':	742 obs. of  7 variables:
#>   .. .. ..$ taxon_ancillary_id: chr [1:742] "1" "2" "3" "4" ...
#>   .. .. ..$ taxon_id          : chr [1:742] "1" "1" "1" "1" ...
#>   .. .. ..$ datetime          : Date[1:742], format: NA ...
#>   .. .. ..$ variable_name     : chr [1:742] "subfamily" "hl" "rel" "rll" ...
#>   .. .. ..$ value             : chr [1:742] "Myrmicinae" "1.1582" "0.172681748" "1.323778277" ...
#>   .. .. ..$ unit              : chr [1:742] NA "millimeter" "millimeter" "millimeter" ...
#>   .. .. ..$ author            : chr [1:742] NA NA NA NA ...
#>   .. ..$ observation_ancillary:'data.frame':	8793 obs. of  5 variables:
#>   .. .. ..$ observation_ancillary_id: chr [1:8793] "1" "2" "3" "4" ...
#>   .. .. ..$ observation_id          : chr [1:8793] "1" "1" "1" "2" ...
#>   .. .. ..$ variable_name           : chr [1:8793] "trap.type" "trap.num" "moose.cage" "trap.type" ...
#>   .. .. ..$ value                   : chr [1:8793] "bait" "1 hour" NA "bait" ...
#>   .. .. ..$ unit                    : chr [1:8793] NA NA NA NA ...
#>   .. ..$ dataset_summary      :'data.frame':	1 obs. of  7 variables:
#>   .. .. ..$ package_id                 : chr "edi.193.4"
#>   .. .. ..$ original_package_id        : chr "knb-lter-hfr.118.32"
#>   .. .. ..$ length_of_survey_years     : num 15
#>   .. .. ..$ number_of_years_sampled    : num 14
#>   .. .. ..$ std_dev_interval_betw_years: num 0.67
#>   .. .. ..$ max_num_taxa               : num 53
#>   .. .. ..$ geo_extent_bounding_box_m2 : num 913451
#>   .. ..$ variable_mapping     :'data.frame':	19 obs. of  6 variables:
#>   .. .. ..$ variable_mapping_id: chr [1:19] "1" "2" "3" "4" ...
#>   .. .. ..$ table_name         : chr [1:19] "observation" "observation_ancillary" "observation_ancillary" "observation_ancillary" ...
#>   .. .. ..$ variable_name      : chr [1:19] "abundance" "trap.type" "trap.num" "moose.cage" ...
#>   .. .. ..$ mapped_system      : chr [1:19] "Darwin Core" "The Ecosystem Ontology" NA NA ...
#>   .. .. ..$ mapped_id          : chr [1:19] "http://rs.tdwg.org/dwc/terms/individualCount" "http://purl.dataone.org/odo/ECSO_00001591" NA NA ...
#>   .. .. ..$ mapped_label       : chr [1:19] "individualCount" "type of trap" NA NA ...

# Clean up
unlink(mypath, recursive = TRUE)