Save an ecocomDP dataset

Save an ecocomDP dataset

save_data(dataset, path, type = ".rds", name = NULL)

Arguments

dataset
(list) One or more datasets of the structure returned by read_data(). Name of the dataset object will become the file name if name is not used.

path
(character) Path to the directory in which dataset will be written.

type
(character) Type of file to save the dataset as. Default is ".rds" but can also be ".csv". Note: metadata and validation_issues are lost when using ".csv".

name
(character) An optional argument for setting the saved file name (for .rds) if you'd like it to be different than dataset's object name.

Value

.rds

If type = ".rds", then an .rds representation of dataset is returned.

.csv

If type = ".csv", then an set of .csv files are written to a sub-directory of path named after the data package/product ID.

Note

Subsequent calls won't overwrite files or directories

Examples

# Create a list of datasets
datasets <- c(ants_L1, ants_L1, ants_L1)  # 3 of the same, with different names
names(datasets) <- c("ds1", "ds2", "ds3")

# Create directory for the data
mypath <- paste0(tempdir(), "/data")
dir.create(mypath)

# Save as .rds
save_data(datasets, mypath)

#> Writing datasets.rds to C:\Users\Colin\AppData\Local\Temp\RtmpYX6pOc/data

dir(mypath)

#> [1] "datasets.rds"

# Save as .rds with the name "mydata"
save_data(datasets, mypath, name = "mydata")

#> Writing mydata.rds to C:\Users\Colin\AppData\Local\Temp\RtmpYX6pOc/data

dir(mypath)

#> [1] "datasets.rds" "mydata.rds"  

# Save as .csv
save_data(datasets, mypath, type = ".csv")

#> Writing ds1, ds2, ds3 as .csv to C:\Users\Colin\AppData\Local\Temp\RtmpYX6pOc/data

dir(mypath)

#> [1] "datasets.rds" "ds1"          "ds2"          "ds3"          "mydata.rds"  

# Clean up
unlink(mypath, recursive = TRUE)